refactor: adopt modern modules rust 2018+
