+++ app/src/application.rs
@@ -1,6 +1,5 @@
-pub mod blockchain;
-pub mod config;
 pub mod debank;
 pub mod exchange;
+pub mod hold;
 pub mod price;
 pub mod sheets;
