+++ app/src/application/exchange.rs
@@ -1,4 +1,4 @@
 pub mod binance;
-pub mod exchange;
 pub mod exchange_balances_routine;
 pub mod kraken;
+pub mod use_cases;
