+++ app/src/application/price.rs
@@ -1,1 +1,2 @@
 pub mod token_prices;
+
