+++ app/src/infrastructure.rs
@@ -1,6 +1,6 @@
 pub mod blockchain;
+pub mod config;
 pub mod debank;
 pub mod exchange;
-pub mod hold;
 pub mod price;
 pub mod sheets;
