+++ app/src/infrastructure/blockchain.rs
@@ -1,4 +1,3 @@
 pub mod chains;
-pub mod explorer;
 pub mod explorers;
 pub mod token;
