+++ app/src/infrastructure/blockchain/token.rs
@@ -1,0 +1,1 @@
+pub mod spam_filter;
