--- app/src/infrastructure/blockchain/token/mod.rs
@@ -1,2 +1,0 @@
-pub mod spam_filter;
-pub mod special_tokens;
