--- app/src/prettyprint/mod.rs
@@ -1,1 +1,0 @@
-pub mod prettyprint;
